Refactor benchmarking code
