+++ backend/Cargo.toml
@@ -10,6 +10,7 @@ futures = "0.3.13"
 futures-timer = "3.0.2"
 glib = { git = "https://github.com/pop-os/gtk-rs" }
 hidapi = { version = "1.2", default-features = false, features = ["linux-static-hidraw"] }
+libc = "0.2"
 once_cell = "1.4"
 ordered-float = { version = "2.0", features = ["serde"] }
 palette = "0.5"
@@ -24,9 +25,6 @@ features = ["hidapi", "std"]
 
 [target.'cfg(target_os = "linux")'.dependencies]
 zbus = "1.9.1"
-
-[dev-dependencies]
-libc = "0.2"
 
 [features]
 appimage = []
